vit-bartpho/requirements.txt
transformers==4.45.2
sentence-transformers==3.1.1
rouge_score
evaluate
underthesea
pycocoevalcap
torch
pillow
pandas
numpy
wandb
datasets
tqdm
matplotlib
